Move functions dedicated to aws config manipulations to separate file
